q_link_sim/requirements.txt
streamlit==1.36.0
numpy==2.2.0
pandas==2.2.2
networkx==3.3
matplotlib==3.8.4
plotly==5.22.0
pytest==8.2.2
cryptography==42.0.5
SQLAlchemy==2.0.31
jsonpickle==3.0.2
web3==6.20.0
eth-account==0.12.1  # Versión compatible con web3==6.20.0
python-dotenv==1.0.1
requests==2.32.3
reportlab==4.2.0
paho-mqtt==1.6.1y